remove unneeded usage of removeReservedFields roll it back to Message folder
